Add files via upload
Removed the ability to press backspace to go to a previous room. The way to proceed to the tavern or mansion is to pick the zombie option when approaching the town. Will implement more states later like a death screen. Might include the possibility in the near future of having the player roll a dice to attack monsters and decide if they win or lose.